+++ src/content/components/DictPanel/index.tsx
@@ -27,6 +27,7 @@ type ChildrenProps =
   >
 
 export interface DictPanelProps extends ChildrenProps {
+  readonly isAnimation: boolean
   readonly dictionaries: DictionariesState['dictionaries']
   readonly allDictsConfig: DictConfigs
   readonly langCode: AppConfig['langCode']
@@ -36,6 +37,7 @@ export interface DictPanelProps extends ChildrenProps {
 export default class DictPanel extends React.Component<DictPanelProps> {
   render () {
     const {
+      isAnimation,
       isFav,
       isPinned,
       langCode,
@@ -53,7 +55,6 @@ export default class DictPanel extends React.Component<DictPanelProps> {
       allDictsConfig,
       panelWidth,
       fontSize,
-      panelDbSearch,
 
       updateItemHeight,
     } = this.props
@@ -64,7 +65,7 @@ export default class DictPanel extends React.Component<DictPanelProps> {
     } = dictionaries
 
     return (
-      <>
+      <div className={`panel-Root${isAnimation ? ' isAnimate' : ''}`}>
         {React.createElement(MenuBar, {
           isFav,
           isPinned,
@@ -90,7 +91,6 @@ export default class DictPanel extends React.Component<DictPanelProps> {
               text: (dictionaries.searchHistory[0] || selection.selectionInfo).text,
               dictURL,
               fontSize,
-              panelDbSearch,
               preferredHeight: allDictsConfig[id].preferredHeight,
               panelWidth,
               searchStatus: (dictsInfo[id] as any).searchStatus,
@@ -100,7 +100,7 @@ export default class DictPanel extends React.Component<DictPanelProps> {
             })
           })}
         </div>
-      </>
+      </div>
     )
   }
 }