+++ src/popup/index.ts
@@ -4,11 +4,11 @@ import VueQriously from 'vue-qriously'
 import VueI18Next from '@panter/vue-i18next'
 import i18nLoader from '@/_helpers/i18n'
 import popupLocles from '@/_locales/popup'
+import { injectSaladictInternal } from '@/_helpers/injectSaladictInternal'
 
-// keep search history
-// window.__SALADICT_INTERNAL_PAGE__ = true
+window.__SALADICT_INTERNAL_PAGE__ = true
 window.__SALADICT_POPUP_PAGE__ = true
-injectPanel() // inject panel AFTER flags are set
+injectSaladictInternal(true) // inject panel AFTER flags are set
 
 Vue.use(VueQriously)
 Vue.use(VueI18Next)
@@ -23,11 +23,3 @@ new Vue({ // eslint-disable-line no-new
   i18n,
   render: h => h(App),
 })
-
-function injectPanel () {
-  const $script = document.createElement('script')
-  $script.src = './content.js'
-  $script.type = 'text/javascript'
-
-  document.body.appendChild($script)
-}